Move chartoption creation into separate hook
